agentic-trader/requirements.txt
langchain 
openai 
groq 
langchain-core 
langchain-community
langchain-openai
langchain-groq
kiteconnect 
python-dotenv
pandas
ta-lib-everywhere
nsepython
flask